Update IAutoRepairTokenRangeSplitter API to be more flexible
Updates IAutoRepairTokenRangeSplitter API to take in a set of tables
instead of a single table and return "RepairAssignments".

This allows the API more control over how the repairs are issued by
the AutoRepair framework, by allowing grouping of repairs by specific
tables or a single repair at the keyspace level.

Also updates IAutoRepairTokenRangeSplitter to be a ParameterizedClass,
so custom splitters can have their own configuration.

Update AutoRepair to account for these changes.

Update DefaultAutoRepairTokenSplitter to account for the change, and
to also evaluate on getRepairByKeyspace option.  This also improves
the filtering of tables that should not be repaired where previously
getRepairByKeyspace would not factor this in.

Move logic to split token ranges evenly into AutoRepairUtils so it
can be reutilized by other splitters.

Adds prototype UnrepairedBytesBasedTokenRange that demonstrates
consuming splitter-based properties to limit the repair assignments
returned to be bound.